controller.js: added print statement to display on the console for debugging purposes.
styles.js: added styles for the login window and the "Welcome, LoggedUser!" that now appears when someone logs in

index.jsp: linked styles.css to it

login.jsp: added classes to text fields, login button, and divs for styling.

main.jsp: added <p> tag that will hold the name of the person that is currently logged in.

dom_functions.js: function getName() that sends the user names of the person logged in and the person clicked on websocket.js for server processing.

websocket.js: implemented sendNewChat() function to that sends both usernames to the server to process and created conditional statement to handle the response. Right now it is only printing the json object in the console window of the browser.